Patched Selectable Points
Added a quick hack to bring selectable data points above all other SVG
elements in the graph. This quick fix currently requires that the data
bound to each group element be rebound to each .datapoint element in
the SVG (.datapoints are no longer contained in the group). This seems
inefficient and should be addressed in future updates.